WIP: tuning curve
for loops are now correct and the good K and E are passed to the SRW
wrapper functions. COde is very slow and must be paralel to really be
used. We cannot use the native paralisation from brac4sr because we send
single energy points. paralel loop must be with the tuning curve
function.

TODO: implement Kim's on axis flux function so tuning curves can be
readily obtained.

 Changes to be committed:
	modified:   barc4sr/aux_processing.py
	modified:   barc4sr/aux_utils.py
	modified:   barc4sr/sr_undulator.py