bug fix: changes to tuning curve
corrections to the tuning curve calculation - model is not yet finished.
It need to be able to calculate the beam in a larger window (containing
the more than the full beam), so that the convolution with the electron
beam phase space works correctly - othewise the photonb flux inside the
slit contains still the single electron flux.

 Changes to be committed:
	modified:   barc4sr/aux_processing.py
	modified:   barc4sr/sr_undulator.py